+++ tests/examples/login.spec.js
@@ -89,7 +89,8 @@ test.afterEach(
                     : undefined,
 
                 {
-                    failureScreenshot
+                    failureScreenshot,
+                    testInfo
                 }
 
             );